iommufd: Add function to find the unique vfio id
When CONFIG_VFIO_DEVICE_CDEV is enabled in the kernel, there will be
vfio device specific information located at
/sys/bus/pci/BUS:DEVICE:FUNCTION/vfio-dev/vfioX; where X is the vfio id
for said device. In this commit we add a function that returns X given a
BUS:DEVICE:FUNCTION id.

This is a preparation commit that adds iommufd functionality in such a
way that we can setup and perform iommufd DMA in subsequent commits.

Signed-off-by: Joel Granados <j.granados@samsung.com>