pci: Get the vfio dev FD using iommufd
We add a function pointer to the fio_device struct that holds call that
will get a vfio device descriptor. It can either be a cdev (iommufd) or
a device found through vfio groups (legacy).
Choose the iommufd function if environment variable VFN_USE_IOMMUFD is
defined. Otherwise stick with the legacy logic.
In this commit we just set up the vfio file descriptor; we still have to
adjust the way DMA is mapped as it is different for legacy vfio an
iommufd.

Signed-off-by: Joel Granados <j.granados@samsung.com>