build: Rename build option to USE_IOMMUFD_BACKEND
This is a preparation commit for when the selection of the iommufd
back-end also depends on user defined configuration instead of just the
presence of a variable in a kernel header file.

When USE_IOMMUFD_BACKEND is true, libvfn is compiled against the kernels
iommufd infra. If it is false, it will be compiled against the legacy
vfio infra.

Signed-off-by: Joel Granados <j.granados@samsung.com>