build: Add force-vfio meson option
Two address space back-ends will coexist for the time being: iommufd and
legacy vfio. Previously we used the iommufd back-end when the kernel
headers defined VFIO_DEVICE_BIND_IOMMUFD and fell back to legacy vfio
when that variable was not defined. This disallowed testing both
back-ends with one kernel. To address this, we add the "force-vfio" user
configuration option to force the use of the legacy vfio back-end when
compiling under kernel header defining VFIO_DEVICE_BIND_IOMMUFD.

After this change the kernel's iommufd back-end is used if
VFIO_DEVICE_BIND_IOMMUFD is defined in the headers and the force-vfio
build option is set to false.

Signed-off-by: Joel Granados <j.granados@samsung.com>