github: Add action to test the two iommu back-ends
Two iommu address space back-ends are supported : vfio and iommufd.
github actions use one or the other based on the kernel headers
available on the action distro. This presents two issues: (1) It is
hard to identify an iommu back-end specific failure as the action titles
do not contain what back-end was used. And (2) the two back-ends are
not tested with the same underlying configuration/distribution.

To address these issues we add a iommu back-end build test that will test
both iommufd and vfio back-ends on top of fedora rawhide. The back-end
will be chosen by passing the force-vfio argument in the meson setup
step.

Signed-off-by: Joel Granados <j.granados@samsung.com>