iommu/iommufd: fix behavior wrt. multiple ioas
With iommufd the struct vfio_container identifies an IOAS. Currently, it
also holds the file descriptor from opening /dev/iommu. Since we only
need to open it onces, store the fd in a global variable instead.

This allows us to properly repurpose vfio_init_container to create an
IOAS and remove it from get_device_fd().

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>