iommu/context: Check for cdev when getting iommu ctx
In order to use iommufd the kernel needs to be configured with
CONFIG_VFIO_DEVICE_CDEV=y. Making sure this is the case at runtime is
not obvious. We can't rely on /proc/config.gz because the kernel might
not have been built with that option. We also can't rely on checking for
/dev/vfio/devices before main is executed because there is a case where
it is not present even though the kernel was compiled with
CONFIG_VFIO_DEVICE_CDEV=y (when all the devices are bound to non vfio
drivers).

Remove the check for /dev/vfio/devices from the "constructor" attribute
(before main) and call it when the iommu ctx is being created. This will
give the user the opportunity to bind the device to vfio thereby making
/dev/vfio/device available to the check.

The actual check is only called once and the result is placed in a
global static variable that will be returned from then on.

Signed-off-by: Joel Granados <j.granados@samsung.com>